+++ subs2cia/CardExport.py
@@ -8,14 +8,17 @@ from pathlib import Path
 import logging
 import tqdm
 import pandas as pd
-
+import unicodedata as ud
+from collections import defaultdict
 
 class CardExport(Common):
     def __init__(self, sources: List[AVSFile], outdir: Path, outstem: Union[str, None], condensed_video: bool, padding: int,
                  demux_overwrite_existing: bool, overwrite_existing_generated: bool,
                  keep_temporaries: bool, target_lang: str, out_audioext: str,
-                 use_all_subs: bool, subtitle_regex_filter: str, subtitle_regex_substrfilter: str,
-                 subtitle_regex_substrfilter_nokeep: bool, audio_stream_index: int, subtitle_stream_index: int,
+                 use_all_subs: bool, subtitle_regex_filter: str,
+                 # subtitle_regex_substrfilter: str,
+                 # subtitle_regex_substrfilter_nokeep: bool,
+                 audio_stream_index: int, subtitle_stream_index: int,
                  ignore_range: Union[List[List[int]], None], ignore_chapters: Union[List[str], None],
                  bitrate: Union[int, None], mono_channel: bool, interactive: bool, normalize_audio: bool, out_audiocodec: str):
         super(CardExport, self).__init__(
@@ -31,8 +34,8 @@ class CardExport(Common):
             out_audioext=out_audioext,
             use_all_subs=use_all_subs,
             subtitle_regex_filter=subtitle_regex_filter,
-            subtitle_regex_substrfilter=subtitle_regex_substrfilter,
-            subtitle_regex_substrfilter_nokeep=subtitle_regex_substrfilter_nokeep,
+            # subtitle_regex_substrfilter=subtitle_regex_substrfilter,
+            # subtitle_regex_substrfilter_nokeep=subtitle_regex_substrfilter_nokeep,
             audio_stream_index=audio_stream_index,
             subtitle_stream_index=subtitle_stream_index,
             ignore_range=ignore_range,
@@ -75,7 +78,8 @@ class CardExport(Common):
             subdata = subtools.SubtitleManipulator(subfile.filepath,
                                                    threshold=0, padding=self.padding,
                                                    ignore_range=ignore_range, audio_length=audiolength)
-            subdata.load(include_all=self.use_all_subs, regex=self.subtitle_regex_filter)
+            subdata.load(include_all=self.use_all_subs, regex=self.subtitle_regex_filter,
+                         substrreplace_regex='', substrreplace_nokeepchanges=False)
             if subdata.ssadata is None:
                 logging.warning(f"Problem loading subtitle data from {self.picked_streams[k]}")
                 self.picked_streams[k] = None
@@ -85,10 +89,13 @@ class CardExport(Common):
 
     def export(self):
         # expose these as options at some point
-        export_audio = True  # will need to rename these if they are going to become cli switches so they don't conflict
-        export_screenshot = True
+        # will need to rename these if they are going to become cli switches so they don't conflict
+        forbidden_chars = ['[' , ']' , '<' , '>' , ':' , '"' , '/' , '?' , '*' , '^' , '\\' , '|']  # see fn disallowed_char in https://github.com/ankitects/anki/blob/main/rslib/src/media/files.rs
+        forbidden_chars = {ord(c): '' for c in forbidden_chars}
+        export_audio = True if self.picked_streams['audio'] is not None else False
+        export_screenshot = True if self.picked_streams['video'] is not None else False
         export_video = False
-        lbda = 0.0
+        lbda = 0.0  # where to get screenshot. 0=start, 1=end, 0.5=middle
         w = -1
         h = -1
         columns = ['text', 'timestamps', 'audioclip', 'screenclip', 'videoclip', 'sources']
@@ -103,28 +110,31 @@ class CardExport(Common):
                    'audioclip': None,
                    'screenclip': None,
                    'videoclip': None,
-                   'sources': f"{self.sources}"[1:-1]}
+                   'sources': ",".join([s.filepath.name for s in self.sources])}
             if export_audio:
-                outpath = self.outdir / (self.outstem + f"({group.group_range[0]}-{group.group_range[1]}).mp3")
+                outpath = self.outdir / (ud.normalize('NFC', self.outstem).translate(forbidden_chars) + f"_{group.group_range[0]}-{group.group_range[1]}.mp3")
                 row['audioclip'] = f"[sound:{outpath.name}]"
-                ffmpeg_trim_audio_clip_atrim_encode(input_file=self.picked_streams['audio'].demux_file,
+                ffmpeg_trim_audio_clip_atrim_encode(input_file=self.picked_streams['audio'].demux_file.filepath,
                                                     stream_index=0,
                                                     timestamp_start=group.group_range[0],
                                                     timestamp_end=group.group_range[1], quality=self.quality,
                                                     to_mono=self.to_mono, normalize_audio=self.normalize_audio,
                                                     outpath=outpath)
             if export_screenshot:
-                outpath = self.outdir / (self.outstem + f"[{group.group_range[0]}-{group.group_range[1]}].jpg")
-                row['screenclip'] = f"<img src='{outpath.name}' />"
+                outpath = self.outdir / (ud.normalize('NFC', self.outstem).translate(forbidden_chars) + f"_{group.group_range[0]}-{group.group_range[1]}.jpg")
+                row['screenclip'] = f"<img src='{outpath.name}'>"
                 timestamp = (1-lbda) * group.group_range[0] + lbda * group.group_range[1]
-                ffmpeg_get_frame_fast(self.picked_streams['video'].file,
+                ffmpeg_get_frame_fast(self.picked_streams['video'].file.filepath,
                                       timestamp=timestamp, outpath=outpath, w=w, h=h)
             if export_video:
-                outpath = self.outdir / (self.outstem + f"_{group.group_range[0]}_{group.group_range[1]}.mp4")
+                outpath = self.outdir / (ud.normalize('NFC', self.outstem).translate(forbidden_chars) + f"_{group.group_range[0]}-{group.group_range[1]}.mp4")
                 row['videoclip'] = f"[sound:{outpath.name}]"
-                ffmpeg_trim_video_clip_directcopy(self.picked_streams['video'].file, timestamp_start=group.group_range[0],
+                ffmpeg_trim_video_clip_directcopy(self.picked_streams['video'].file.filepath,
+                                                  timestamp_start=group.group_range[0],
                                                     timestamp_end=group.group_range[1], quality=None, outpath=outpath)
             exported = exported.append([row], ignore_index=True)
+            if exported.shape[0] > 10:  # DEBUG ONLY
+                break
         # print(exported)
         outpath = self.outdir / (self.outstem + ".tsv")
-        exported.to_csv(outpath, sep='\t')
+        exported.to_csv(outpath, sep='\t', index=False, header=False)